# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Universal POS API Documentation",
        "description": "API documentation for the POS system",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/sales": {
            "get": {
                "tags": [
                    "Sales"
                ],
                "summary": "Get sales",
                "operationId": "248121b133c730c6860e94383dfe3e12",
                "responses": {
                    "200": {
                        "description": "Sales retrieved"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Sales"
                ],
                "summary": "Create sale",
                "operationId": "42cd0cb8dcc906da8e52634de671f7bf",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Sale created"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/sales/{id}": {
            "get": {
                "tags": [
                    "Sales"
                ],
                "summary": "Get sale details",
                "operationId": "b8c1920044f4a939f5bd289d5693854a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sale retrieved"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/sales/{id}/refund": {
            "post": {
                "tags": [
                    "Sales"
                ],
                "summary": "Refund a sale",
                "operationId": "280a4266f19502c1275bc1e638162c0e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Refunded"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/services": {
            "get": {
                "tags": [
                    "Services"
                ],
                "summary": "Get services",
                "operationId": "97de511018def32ca0ccbc800756c538",
                "responses": {
                    "200": {
                        "description": "Services retrieved"
                    }
                }
            },
            "post": {
                "tags": [
                    "Services"
                ],
                "summary": "Create a service",
                "operationId": "0634f2b3ac65bd2815f4b6459f0db97c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "slug"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Service created"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/services/{id}": {
            "put": {
                "tags": [
                    "Services"
                ],
                "summary": "Update service",
                "operationId": "14e625e6fca9956d1143de1f41a8a228",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Service updated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Services"
                ],
                "summary": "Delete service",
                "operationId": "9f75ff1ee1d79f882f516173057387ec",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/service-transactions": {
            "get": {
                "tags": [
                    "Services"
                ],
                "summary": "Get transactions",
                "operationId": "a016e197840dfc6ee4af7a794d36f50e",
                "responses": {
                    "200": {
                        "description": "Transactions retrieved"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Services"
                ],
                "summary": "Create transaction",
                "operationId": "4ed7b75c9db4edd1049846416e63c0ed",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "service_id",
                                    "amount"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}